requirements.txt
djangorestframework
psycopg2
pyyaml
uritemplate
